llama-index-integrations/vector_stores/llama-index-vector-stores-mongodb/pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.codespell]
check-filenames = true
check-hidden = true
skip = "*.csv,*.html,*.json,*.jsonl,*.pdf,*.txt,*.ipynb"

[tool.llamahub]
contains_example = false
import_path = "llama_index.vector_stores.mongodb"

[tool.llamahub.class_authors]
MongoDBAtlasVectorSearch = "llama-index"

[tool.mypy]
disallow_untyped_defs = true
exclude = ["_static", "build", "examples", "notebooks", "venv"]
ignore_missing_imports = true
python_version = "3.8"

[tool.poetry]
authors = [
    "The MongoDB Python Team",
]
description = "llama-index vector_stores mongodb integration"
exclude = ["**/BUILD"]
license = "MIT"
name = "llama-index-vector-stores-mongodb"
readme = "README.md"
version = "0.1.5"

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
llama-index-core = "^0.10.1"
pymongo = "^4.6.1"

[tool.poetry.group.dev.dependencies]
ipython = "8.10.0"
jupyter = "^1.0.0"
llama-index-embeddings-openai = "^0.1.5"
llama-index-llms-openai = "^0.1.13"
llama-index-readers-file = "^0.1.4"
mypy = "0.991"
pre-commit = "3.2.0"
pylint = "2.15.10"
pytest = "7.2.1"
pytest-mock = "3.11.1"
ruff = "0.0.292"
tree-sitter-languages = "^1.8.0"
types-Deprecated = ">=0.1.0"
types-PyYAML = "^6.0.12.12"
types-protobuf = "^4.24.0.4"
types-redis = "4.5.5.0"
types-requests = "2.28.11.8"
types-setuptools = "67.1.0.0"

[tool.poetry.group.dev.dependencies.black]
extras = ["jupyter"]
version = "<=23.9.1,>=23.7.0"

[tool.poetry.group.dev.dependencies.codespell]
extras = ["toml"]
version = ">=v2.2.6"

[[tool.poetry.packages]]
include = "llama_index/"

[tool.pytest.ini_options]
markers = [
    "requires_mongodb_uri: mark tests as needing MONGODB_URI in os.environ",
    "requires_openai_key: mark tests as needing OPENAI_API_KEY in os.environ",
]
